+++ hammer-sqldb/src/main/java/cn/featherfly/hammer/sqldb/jdbc/operate/MergeOperate.java
@@ -1,11 +1,13 @@
 package cn.featherfly.hammer.sqldb.jdbc.operate;
 
+import java.sql.PreparedStatement;
 import java.util.Map;
 
 import com.speedment.common.tuple.Tuple2;
 
 import cn.featherfly.common.db.mapping.ClassMappingUtils;
 import cn.featherfly.common.db.metadata.DatabaseMetadata;
+import cn.featherfly.common.lang.ArrayUtils;
 import cn.featherfly.common.repository.mapping.ClassMapping;
 import cn.featherfly.hammer.sqldb.jdbc.Jdbc;
 
@@ -64,7 +66,17 @@ public class MergeOperate<T> extends AbstractOperate<T> {
     public int execute(final T entity, boolean onlyNull) {
         Tuple2<String, Map<Integer, String>> tuple = ClassMappingUtils.getMergeSqlAndParamPositions(entity,
                 classMapping, onlyNull, jdbc.getDialect());
-        return jdbc.update(tuple.get0(), getParameters(entity, tuple.get1()));
+        return jdbc.execute((con, manager) -> {
+            try (PreparedStatement prep = con.prepareStatement(tuple.get0())) {
+                Object[] params = setParameters(entity, tuple.get1(), prep, manager);
+                if (logger.isDebugEnabled()) {
+                    logger.debug("execute sql: {} \n params: {}", sql, ArrayUtils.toString(params));
+                }
+                int result = prep.executeUpdate();
+                return result;
+            }
+        });
+        //        return jdbc.update(tuple.get0(), getParameters(entity, tuple.get1()));
     }
 
     /**