+++ hammer-sqldb/change-log.md
@@ -1,6 +1,7 @@
 # 0.5.6 
     1.Jdbc方法参数位置修改
     2.修复JdbcImpl没有释放connection的问题
+    3.Asm使用ClassLoader加载类信息，可以加载fatjar模式的类
 
 # 0.5.5 2020-12-8
     1.DSL条件查询加入group(Fucntion<C,L> function)方法 ，不用手动调用endGroup方法